Improving Password Manager with validation and efficiency improvements.
